lambda done

+++ pkg/secrets/awsstore/aws.go
@@ -107,6 +107,7 @@ func (s *AWSSecretsStore) RetrieveOauthState(state string) (string, error) {
 	//get full name including app name and environment type
 	fullParamName := fmt.Sprintf("/%s/%s", s.StorePrefix, state)
 
+	s.Log.Tracef("Retrieving state %s", fullParamName)
 	input := &ssm.GetParameterInput{
 		Name:           aws.String(fullParamName),
 		WithDecryption: aws.Bool(false),
@@ -117,11 +118,23 @@ func (s *AWSSecretsStore) RetrieveOauthState(state string) (string, error) {
 		return "", err
 	}
 
+	//remove quotes
+	var resultString string
+	resultRaw := *result.Parameter.Value
+	//check if string is quoted and strip if yes
+	if strings.HasPrefix(resultRaw, "\"") {
+		resultString = strings.Trim(resultRaw, "\"")
+		s.Log.Tracef("Retrieved %s", resultString)
+		return resultString, nil
+	}
+	resultString = resultRaw
+
 	value := *result.Parameter.Value
 
 	//clean up. Delete retrieved state
+	s.Log.Tracef("Deleting state %s", fullParamName)
 	delInput := &ssm.DeleteParameterInput{
-		Name: aws.String(state),
+		Name: aws.String(fullParamName),
 	}
 
 	_, err = ssmSvc.DeleteParameter(delInput)
@@ -163,7 +176,7 @@ func (s *AWSSecretsStore) RetrieveVar(varName string) (varValue string, err erro
 	return resultString, nil
 }
 
-func Initialize(storePrefix string, awsRegion string) (*AWSSecretsStore, error) {
+func Initialize(log *logrus.Logger, storePrefix string, awsRegion string) (*AWSSecretsStore, error) {
 	accessTokenParamName := fmt.Sprintf("/%s/access_token", storePrefix)
 	refreshTokenParamName := fmt.Sprintf("/%s/refresh_token", storePrefix)
 
@@ -179,7 +192,9 @@ func Initialize(storePrefix string, awsRegion string) (*AWSSecretsStore, error) 
 	return &AWSSecretsStore{AccessTokenParamName: accessTokenParamName,
 		RefreshTokenParamName: refreshTokenParamName,
 		AWSSession:            sess,
-		StorePrefix:           storePrefix}, nil
+		StorePrefix:           storePrefix,
+		Log:                   log,
+	}, nil
 }
 
 func (s *AWSSecretsStore) Close() error {
